// frontend/src/lib/api/models/GalleryImageResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GalleryImageResponseDto = {
    id: string;
    albumId: string;
    filePath: string;
    caption?: string | null;
    sortOrder: number;
    isPublished: boolean;
    createdAt: string;
    updatedAt?: string | null;
    /**
     * URL pública de la imagen
     */
    url?: string;
};